delineate variable parsing and checking
delineates the variable parsing and checking steps so that, for MLR, the checking steps are bypassed.

this means that, in the SLR context, infer will error out (where it did previously) for checks that aren’t necessary if the object is just being passed to `fit()`. i think this is okay, as:

1. i think it’s important that an error in the specify step actually come about at the specify step, for debuggability

2. that extra information will still be able to be able to be passed to modeling functions as preprocessing steps