define a custom `+.gg` method for infer layers
classes shade_* output as `infer_layer` makes use of `ggplot_add` to define a custom `+.gg` layer.

a method for the `+` operator for infer objects.
- object to add (arguments to the RHS of the `+`)
- plot is the existing plot (on the LHS of the `+`)
- object_name is the name of the function (called on the RHS of the `+`)

output is the actual output of the addition - this allows for a more %>%-esque programming style

the biggest advantage this offers us is that we can overwrite existing elements, i.e. subsetting into the patchwork, modifying its elements (for p-value and confidence interval shading), and then overwriting them.