bring attrs along in `infer_layer`s
since the `object_name` in `+.gg` is actually the output of `deparse(substitute(call))`, if that output of a `shade_*` function was stored in an object, the layer addition would no longer work. instead, store each of the arguments as attributes in the `infer_layer` object and then build the `*_term` call out of them.

also reverts the output of visualizations from `calculate`-based workflows to `gg` objects--i'd anticipate this would be a somewhat breaking change if one couldn't add layers to visualize output like other ggplot objects. this means that output is internally inconsistent (though it already is in `calculate` vs `fit` based workflows, which is fine)