add infrastructure for causal inf
As-is, the package only carries around variables identified as outcomes or predictors. This makes bringing the actual outcome (as in, not propensity) variable along during the propensity modeling stage a bit tricky. The causal branch implements a keep argument to specify() which "unused" variables to bring along.

There is already an issue on the package repository proposing an argument, called summarize here, that summarizes fit.infer() output given some function other than broom::tidy(). Supplying something like broom::augment() to that argument effectively links the propensity model and the outcome model in one infer "pipeline."